Combine keeper and remove deprecated function in utils/svcmode
