+++ ailurus/svcmodes/sample/config.json
@@ -1,3 +1,4 @@
 {
-    "display": "Sample Mode"
+    "display": "Sample Mode",
+    "doc": "This service mode is part of the unit test."
 }
