+++ ailurus/utils/svcmode.py
@@ -1,4 +1,5 @@
 from importlib.machinery import SourceFileLoader
+from types import ModuleType
 import ailurus.svcmodes
 import os
 
@@ -6,7 +7,10 @@ def get_svcmode_module(service_mode: str):
     svcmode_dir = os.path.dirname(ailurus.svcmodes.__file__)
     package_dir = os.path.join(svcmode_dir, service_mode)
         
-    return SourceFileLoader(
+    loader = SourceFileLoader(
         f"ailurus.svcmodes.{service_mode}",
         os.path.join(package_dir, "__init__.py")
-    ).load_module()
+    )
+    mod = ModuleType(loader.name)
+    loader.exec_module(mod)
+    return mod
